Move the implementations of `CacheUpdate` into the `cache::event` module
It slices off 26% of the `model/event.rs` file.
